# frontend/generated/api/openapi.json
{
  "components": {
    "schemas": {
...
      "ManualSessionCreate": {
        "additionalProperties": false,
        "description": "A session the athlete performed and is typing in (B-6).\n\nNo file, no recording row, no streams \u2014 a session row with\n``recording_kind=manual`` and its sets. Strength by default, because that\nis what has no device file worth ingesting.",
        "properties": {
          "discipline": {
            "$ref": "#/components/schemas/SessionDiscipline",
            "default": "strength"
          },
          "duration_s": {
            "maximum": 86400.0,
            "minimum": 60.0,
            "title": "Duration S",
            "type": "integer"
          },
          "notes": {
            "anyOf": [
              {
                "maxLength": 4000,
                "pattern": "^[^\\x00]*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "rpe": {
            "anyOf": [
              {
                "maximum": 10.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rpe"
          },
          "sets": {
            "items": {
              "$ref": "#/components/schemas/LoggedSetCreate"
            },
            "maxItems": 200,
            "title": "Sets",
            "type": "array"
          },
          "start_time": {
            "format": "date-time",
            "title": "Start Time",
            "type": "string"
          },
          "timezone": {
            "default": "UTC",
            "description": "IANA name, fixed offset (UTC+02:00), or UTC.",
            "maxLength": 64,
            "pattern": "^[^\\x00]*$",
            "title": "Timezone",
            "type": "string"
          }
        },
        "required": ["start_time", "duration_s"],
        "title": "ManualSessionCreate",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "APIKeyCookie": {
        "in": "cookie",
        "name": "arc_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "title": "arc",
    "version": "0.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/manual-sessions": {
      "post": {
        "description": "Record a session performed without a device file \u2014 a gym session (B-6).\n\nProduces the same session row a file would, with\n`recording_kind=manual` and no recording: WP-6 matches both with one\nquery, and the sets are stored so WP-7's strength alignment has something\nto read.",
        "operationId": "sessions-create_manual_session",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ManualSessionCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            },
            "description": "Malformed body"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            },
            "description": "No valid session"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            },
            "description": "No such session"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorDetail"
                }
              }
            },
            "description": "The session violates a schema or domain rule"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "summary": "Create Manual Session",
        "tags": ["sessions"]
      }
    },
    "/api/v1/matches": {
      "get": {
        "description": "List match links, newest first. `?status=pending` is the proposal inbox.\n\nEvery row carries both sides, so the inbox renders and is answered without\na request per proposal.",
        "operationId": "matches-list_matches",
        "parameters": [
          {
            "description": "Restrict to one link status; omit for all of them. The proposal inbox is `pending`.",
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/MatchLinkStatus",
              "description": "Restrict to one link status; omit for all of them. The proposal inbox is `pending`."
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "maximum": 2147483647,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_MatchRead_"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            },
            "description": "No valid session"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "summary": "List Matches",
        "tags": ["matches"]
      },
      "post": {
        "description": "Link a session to a planned session by hand (build plan WP-6.6).\n\nAlways the athlete's own, and therefore **sticky**: no re-run of matching\nrevises or removes it. With `displaced` set it is the executed-instead-of\nlink \u2014 the planned session becomes `displaced` rather than completed, and\nthe activity is scored standalone.\n\nThe similarity is still computed and stored, however low: a deliberate link\nat 0.12 is worth being able to look at afterwards.",
        "operationId": "matches-create_match",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MatchCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatchRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            },
            "description": "Malformed body"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            },
            "description": "No valid session"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            },
            "description": "No such session"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            },
            "description": "Already linked, or already ruled on"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorDetail"
                }
              }
            },
            "description": "The link violates a schema or domain rule"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "summary": "Create Match",
        "tags": ["matches"]
      }
    },
...
    "/api/v1/sessions": {
      "get": {
        "description": "List completed sessions, newest first, optionally within a date range.\n\nA log rather than a calendar, so it reads backwards \u2014 the opposite of\n`GET /planned-sessions`, and deliberately: what happened is read from the\nmost recent, what is planned is read forwards.",
        "operationId": "sessions-list_sessions",
        "parameters": [
          {
            "description": "Earliest athlete-local date to include (inclusive).",
            "in": "query",
            "name": "start",
            "required": false,
            "schema": {
              "description": "Earliest athlete-local date to include (inclusive).",
              "format": "date",
              "title": "Start",
              "type": "string"
            }
          },
          {
            "description": "Latest athlete-local date to include (inclusive).",
            "in": "query",
            "name": "end",
            "required": false,
            "schema": {
              "description": "Latest athlete-local date to include (inclusive).",
              "format": "date",
              "title": "End",
              "type": "string"
            }
          },
          {
            "description": "Restrict to one discipline; omit for all of them.",
            "in": "query",
            "name": "discipline",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/SessionDiscipline",
              "description": "Restrict to one discipline; omit for all of them."
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "maximum": 2147483647,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_SessionListItem_"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            },
            "description": "No valid session"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "summary": "List Sessions",
        "tags": ["sessions"]
      }
    },
...
    "/api/v1/sessions/{session_id}/streams": {
      "get": {
        "description": "The per-second samples behind one session, for the charts.\n\nIts own resource because it is 1-2 MB for a long ride (A4.1: 14 400 rows\nper channel for four hours). Every channel is the **cleaned** column with\nits nulls intact \u2014 a recording stop is a break in the trace, not a run of\nzeros \u2014 and the anomaly regions come with it so the chart can mark what\nwas repaired (A4.2).\n\n404 for a session that has no recording: a gym session typed in by hand\nnever had samples, and the detail says so because that sentence is the\nempty state the page renders.",
        "operationId": "sessions-get_session_streams",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionStreamsRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            },
            "description": "No valid session"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            },
            "description": "No such session"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "summary": "Get Session Streams",
        "tags": ["sessions"]
      }
    },
...
  }
}